@@ -1,42 +1,41 @@
-﻿namespace GenshinToolbox.ArtScraper
+﻿namespace GenshinToolbox.ArtScraper;
+
+enum ArtSet
 {
-	enum ArtSet
-	{
-		Adventurer,
-		ArchaicPetra,
-		Berserker,
-		BlizzardStrayer,
-		BloodstainedChivalry,
-		BraveHeart,
-		CrimsonWitchOfFlames,
-		DefendersWill,
-		EmblemOfSeveredFate,
-		Gambler,
-		GladiatorsFinale,
-		HeartOfDepth,
-		Initiate,
-		Instructor,
-		Lavawalker,
-		LuckyDog,
-		MaidenBeloved,
-		MartialArtist,
-		NoblesseOblige,
-		PaleFlame,
-		PrayersForDestiny,
-		PrayersForIllumination,
-		PrayersForWisdom,
-		PrayersToSpringtime,
-		ResolutionOfSojourner,
-		RetracingBolide,
-		Scholar,
-		ShimenawasReminiscence,
-		TenacityOfTheMillelith,
-		TheExile,
-		ThunderingFury,
-		Thundersoother,
-		TinyMiracle,
-		TravelingDoctor,
-		ViridescentVenerer,
-		WanderersTroupe,
-	}
+    Adventurer,
+    ArchaicPetra,
+    Berserker,
+    BlizzardStrayer,
+    BloodstainedChivalry,
+    BraveHeart,
+    CrimsonWitchOfFlames,
+    DefendersWill,
+    EmblemOfSeveredFate,
+    Gambler,
+    GladiatorsFinale,
+    HeartOfDepth,
+    Initiate,
+    Instructor,
+    Lavawalker,
+    LuckyDog,
+    MaidenBeloved,
+    MartialArtist,
+    NoblesseOblige,
+    PaleFlame,
+    PrayersForDestiny,
+    PrayersForIllumination,
+    PrayersForWisdom,
+    PrayersToSpringtime,
+    ResolutionOfSojourner,
+    RetracingBolide,
+    Scholar,
+    ShimenawasReminiscence,
+    TenacityOfTheMillelith,
+    TheExile,
+    ThunderingFury,
+    Thundersoother,
+    TinyMiracle,
+    TravelingDoctor,
+    ViridescentVenerer,
+    WanderersTroupe,
 }
