@@ -1,25 +1,24 @@
-﻿namespace GenshinToolbox.ArtScraper
-{
-	enum Stat
-	{
-		Atk,
-		AtkPerc,
-		Hp,
-		HpPerc,
-		Def,
-		DefPerc,
-		EnergyRecharge,
-		ElementalMastery,
-		CritDmg,
-		CritRate,
+﻿namespace GenshinToolbox.ArtScraper;
 
-		PyroDmgBonus,
-		HydroDmgBonus,
-		ElectroDmgBonus,
-		AnemoDmgBonus,
-		CryoDmgBonus,
-		GeoDmgBonus,
-		PhysicalDmgBonus,
-		HealingBonusPerc,
-	}
+enum Stat
+{
+    Atk,
+    AtkPerc,
+    Hp,
+    HpPerc,
+    Def,
+    DefPerc,
+    EnergyRecharge,
+    ElementalMastery,
+    CritDmg,
+    CritRate,
+
+    PyroDmgBonus,
+    HydroDmgBonus,
+    ElectroDmgBonus,
+    AnemoDmgBonus,
+    CryoDmgBonus,
+    GeoDmgBonus,
+    PhysicalDmgBonus,
+    HealingBonusPerc,
 }
